+++ packages/typegpu/src/data/vector.ts
@@ -1,4 +1,4 @@
-import { createDualImpl } from '../shared/generators.ts';
+import { createDualImpl } from '../core/function/dualImpl.ts';
 import { $repr } from '../shared/symbols.ts';
 import { snip } from './snippet.ts';
 import { bool, f16, f32, i32, u32 } from './numeric.ts';
