+++ packages/typegpu/src/std/discard.ts
@@ -1,6 +1,6 @@
 import { snip } from '../data/snippet.ts';
 import { Void } from '../data/wgslTypes.ts';
-import { createDualImpl } from '../shared/generators.ts';
+import { createDualImpl } from '../core/function/dualImpl.ts';
 
 export const discard = createDualImpl(
   // CPU
