+++ packages/typegpu/src/std/matrix.ts
@@ -1,7 +1,7 @@
 import type { Snippet } from '../data/snippet.ts';
 import { mat4x4f } from '../data/matrix.ts';
 import type { m4x4f, v3f } from '../data/wgslTypes.ts';
-import { createDualImpl } from '../shared/generators.ts';
+import { createDualImpl } from '../core/function/dualImpl.ts';
 import { mul } from './numeric.ts';
 
 /**
